+++ scripts/do-split.sh
@@ -14,6 +14,7 @@ $GIT_SUBSPLIT publish "
     src/React/Stream:git@github.com:reactphp/stream.git
     src/React/Cache:git@github.com:reactphp/cache.git
     src/React/Socket:git@github.com:reactphp/socket.git
+    src/React/SocketClient:git@github.com:reactphp/socket-client.git
     src/React/Http:git@github.com:reactphp/http.git
     src/React/HttpClient:git@github.com:reactphp/http-client.git
     src/React/Dns:git@github.com:reactphp/dns.git