+++ src/React/EventLoop/LoopInterface.php
@@ -2,6 +2,8 @@
 
 namespace React\EventLoop;
 
+use React\EventLoop\Timer\TimerInterface;
+
 interface LoopInterface
 {
     public function addReadStream($stream, $listener);
@@ -13,7 +15,8 @@ interface LoopInterface
 
     public function addTimer($interval, $callback);
     public function addPeriodicTimer($interval, $callback);
-    public function cancelTimer($signature);
+    public function cancelTimer(TimerInterface $timer);
+    public function isTimerActive(TimerInterface $timer);
 
     public function tick();
     public function run();