+++ src/React/SocketClient/Connector.php
@@ -1,18 +1,17 @@
 <?php
 
-namespace React\HttpClient;
+namespace React\SocketClient;
 
 use React\EventLoop\LoopInterface;
 use React\Dns\Resolver\Resolver;
 use React\Stream\Stream;
+use React\Promise\When;
 use React\Promise\Deferred;
-use React\Promise\FulfilledPromise;
-use React\Promise\RejectedPromise;
 
-class ConnectionManager implements ConnectionManagerInterface
+class Connector implements ConnectorInterface
 {
-    protected $loop;
-    protected $resolver;
+    private $loop;
+    private $resolver;
 
     public function __construct(LoopInterface $loop, Resolver $resolver)
     {
@@ -20,25 +19,36 @@ class ConnectionManager implements ConnectionManagerInterface
         $this->resolver = $resolver;
     }
 
-    public function getConnection($host, $port)
+    public function createTcp($host, $port)
     {
         $that = $this;
 
         return $this
             ->resolveHostname($host)
             ->then(function ($address) use ($port, $that) {
-                return $that->getConnectionForAddress($address, $port);
+                return $that->createSocketForAddress($address, $port);
             });
     }
 
-    public function getConnectionForAddress($address, $port)
+    public function createUdp($host, $port)
     {
-        $url = $this->getSocketUrl($address, $port);
+        $that = $this;
 
-        $socket = stream_socket_client($url, $errno, $errstr, ini_get("default_socket_timeout"), STREAM_CLIENT_CONNECT | STREAM_CLIENT_ASYNC_CONNECT);
+        return $this
+            ->resolveHostname($host)
+            ->then(function ($address) use ($port, $that) {
+                return $that->createSocketForAddress($address, $port, 'udp');
+            });
+    }
+
+    public function createSocketForAddress($address, $port, $transport = 'tcp')
+    {
+        $url = $this->getSocketUrl($address, $port, $transport);
+
+        $socket = stream_socket_client($url, $errno, $errstr, 0, STREAM_CLIENT_CONNECT | STREAM_CLIENT_ASYNC_CONNECT);
 
         if (!$socket) {
-            return new RejectedPromise(new \RuntimeException(
+            return When::reject(new \RuntimeException(
                 sprintf("connection to %s:%d failed: %s", $address, $port, $errstr),
                 $errno
             ));
@@ -50,6 +60,7 @@ class ConnectionManager implements ConnectionManagerInterface
 
         return $this
             ->waitForStreamOnce($socket)
+            ->then(array($this, 'checkConnectedSocket'))
             ->then(array($this, 'handleConnectedSocket'));
     }
 
@@ -61,10 +72,22 @@ class ConnectionManager implements ConnectionManagerInterface
 
         $this->loop->addWriteStream($stream, function ($stream) use ($loop, $deferred) {
             $loop->removeWriteStream($stream);
+
             $deferred->resolve($stream);
         });
 
         return $deferred->promise();
+    }
+
+    public function checkConnectedSocket($socket)
+    {
+        // The following hack looks like the only way to
+        // detect connection refused errors with PHP's stream sockets.
+        if (false === stream_socket_get_name($socket, true)) {
+            return When::reject(new ConnectionException('Connection refused'));
+        }
+
+        return When::resolve($socket);
     }
 
     public function handleConnectedSocket($socket)
@@ -72,18 +95,17 @@ class ConnectionManager implements ConnectionManagerInterface
         return new Stream($socket, $this->loop);
     }
 
-    protected function getSocketUrl($host, $port)
+    protected function getSocketUrl($host, $port, $transport)
     {
-        return sprintf('tcp://%s:%s', $host, $port);
+        return sprintf('%s://%s:%s', $transport, $host, $port);
     }
 
     protected function resolveHostname($host)
     {
         if (false !== filter_var($host, FILTER_VALIDATE_IP)) {
-            return new FulfilledPromise($host);
+            return When::resolve($host);
         }
 
         return $this->resolver->resolve($host);
     }
 }
-