+++ src/React/SocketClient/ConnectorInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace React\SocketClient;
+
+interface ConnectorInterface
+{
+    public function createTcp($host, $port);
+    public function createUdp($host, $port);
+}